# feat: Article VII Phase 1 - Constitutional Telemetry