… logs/constitutional_telemetry/events_20251004.jsonl
… tests/test_constitutional_telemetry.py
… tests/test_constitutional_telemetry_models.py